requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.1
sarvam==0.0.0
sarvamai==0.1.19
sniffio==1.3.1
typing-extensions==4.15.0
typing-inspection==0.4.1
websockets==15.0.1
